removed uploadlogic.js and added success alert in videoInput.js
