+++ src/utils/VideoInput.js
@@ -7,6 +7,12 @@ import TagsInput from "../components/uploadLogic/TagsInput";
 
 export default function VideoInput(props) {
   const [tags, setTags] = useState([]);
+  const [source, setSource] = useState();
+  const [file, setFile] = useState(null);
+  const [title, setTitle] = useState("");
+  const [uploading, setUploading] = useState(false);
+  const [uploadSuccess, setUploadSuccess] = useState(false);
+
   const { width, height } = props;
 
   // console.log("UID in SomeOtherComponent:", uid); // Log the value of uid
@@ -19,11 +25,6 @@ export default function VideoInput(props) {
     }
   }, [user, loading, uid]);
   
-  const [source, setSource] = useState();
-  const [file, setFile] = useState(null);
-  const [title, setTitle] = useState("");
-  const [uploading, setUploading] = useState(false);
-
 
   const inputRef = useRef();
 
@@ -70,6 +71,9 @@ export default function VideoInput(props) {
           try {
             await setDoc(doc(firestore, "videos", docRef.id), videoData);
             await setDoc(doc(firestore, `Users/${uid}/userContent`, docRef.id), videoData);
+            setUploadSuccess(true);
+            setTitle("");
+            setFile(null);
           } catch (error) {
             console.log("Error adding document", error);
           }
@@ -82,6 +86,19 @@ export default function VideoInput(props) {
   const handleChoose = () => {
     inputRef.current.click();
   };
+
+  
+  useEffect(() => {
+    if (uploadSuccess) {
+      const timeoutId = setTimeout(() => {
+        setUploadSuccess(false);
+      }, 2000);
+
+      return () => {
+        clearTimeout(timeoutId);
+      };
+    }
+  }, [uploadSuccess]);
 
   return (
     <form onSubmit={handleUpload} className="flex flex-col align-center border-2 p-6 border-yellow-400 rounded-3xl justify-center bg-zinc-200 bg-opacity-20">
@@ -128,6 +145,11 @@ export default function VideoInput(props) {
       )}
       <div className="min-h-fit leading-10 text-center w-full text-white text-opacity-40">{source || "Nothing selected"}</div>
       <button type="submit" disabled={uploading} className="justify-center h-12 px-6  w-full bg-yellow-500 mt-8 rounded font-semibold text-sm text-black hover:bg-yellow-400">Upload</button>
+      {uploadSuccess && (
+        <div className="flex items-center justify-center mt-4 text-green-500">
+          Video uploaded successfully!
+        </div>
+      )}
     </form>
   );
 }
